refactor: move database dir in @core
