+++ src/core/database/flexsearch/FlexSearchIndex.ts
