+++ src/core/database/flexsearch/Index.worker.ts
