+++ src/core/database/flexsearch/index.ts
