+++ src/core/database/sqlite/SQLiteDatabase.worker.ts
