+++ src/core/database/sqlite/SQLiteDatabaseWorker.ts
