+++ src/core/features/integrity/FilesIntegrityController.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment jsdom
 
 import { getUUID } from 'src/__tests__/utils/uuid';
-import { openSQLite } from '@core/storage/database/sqlite/openSQLite';
+import { openSQLite } from '@core/database/sqlite/openSQLite';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 
 import { AttachmentsController } from '../attachments/AttachmentsController';
