+++ src/core/features/notes/bin/DeletedNotesController.test.ts
@@ -1,7 +1,7 @@
 import ms from 'ms';
 import { getUUID } from 'src/__tests__/utils/uuid';
 import { Mock } from 'vitest';
-import { openSQLite } from '@core/storage/database/sqlite/openSQLite';
+import { openSQLite } from '@core/database/sqlite/openSQLite';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 
 import { NotesController } from '../controller/NotesController';
