+++ src/core/features/notes/controller/__tests__/NotesController.bench.ts
@@ -2,12 +2,12 @@ import { getLongText } from 'src/__tests__/samples';
 import { getUUID } from 'src/__tests__/utils/uuid';
 import { bench } from 'vitest';
 import z from 'zod';
+import { openSQLite } from '@core/database/sqlite/openSQLite';
 import { InMemoryFS } from '@core/features/files/InMemoryFS';
 import { StateFile } from '@core/features/files/StateFile';
-import { openSQLite } from '@core/storage/database/sqlite/openSQLite';
 import { createFileControllerMock } from '@utils/mocks/fileControllerMock';
 
-import { FlexSearchIndex } from '../../../index/flexsearch/FlexSearchIndex';
+import { FlexSearchIndex } from '../../../../database/flexsearch/FlexSearchIndex';
 
 import { NotesController } from '../NotesController';
 import { NotesTextIndexer } from '../NotesTextIndexer';
